Get rid of ambiguous longs from CrawlHashTable and a lot of misc places.

+++ crawl-ref/source/shopping.h
@@ -50,7 +50,7 @@ public:
     bool del_thing(const item_def &item, const level_pos* pos = NULL);
     bool del_thing(std::string desc, const level_pos* pos = NULL);
 
-    unsigned int cull_identical_items(const item_def& item, long cost = -1);
+    unsigned int cull_identical_items(const item_def& item, int cost = -1);
 
     void gold_changed(int old_amount, int new_amount);
 
@@ -84,7 +84,7 @@ private:
     static const item_def&   get_thing_item(const CrawlHashTable& thing);
     static       std::string get_thing_desc(const CrawlHashTable& thing);
 
-    static long      thing_cost(const CrawlHashTable& thing);
+    static int       thing_cost(const CrawlHashTable& thing);
     static level_pos thing_pos(const CrawlHashTable& thing);
 
     static std::string name_thing(const CrawlHashTable& thing,
